refactor: extract user schemas
